requirements.txt
streamlit==1.40.2
agno>=2.2.10
Pillow>=10.0.0
duckduckgo-search>=6.0.0
ddgs
groq>=0.9.0
fpdf2
fastapi
uvicorn
python-multipart